requirements.txt
Flask==2.2.2
Flask_WTF==1.0.1
WTForms==3.0.1
email_validator
flask-sqlalchemy
flask-bcrypt
flask_login
pillow
flask-mail
itsdangerous==2.0